Refactor origin initialization to common library
A good chunk of the origin initialization settings are common
(or, will be common) to the other servers that you can start with
Pelican.  Refactor the initialization out so it can be reusable as
was already done for the client.
