+++ config/resources/osdf.yaml
